S7Adapter Update private byte[] ReadBytesRaw(string address, uint bytes) 	//Updated version 1.1,             /* [0..65535]= 65536 Maximum DB bytes in siemens PLC              * Each time we are fetching 219 bytes(maximum =220) into bReceive[] and              * fill into data[] (data.length = bytes)              * (note:testing of this new code is pending)              * Update Author: Saad Liaqat Ali. ver 1.1              * Orignal Author: Kevin Boronka. ver 1.0              * Update Date: April 4,2015; 6:00pm.              */ S7Address Update and i also changed ushort to uint for calculation to get 65536 bytes

+++ sar/S7Siemens/S7Address.cs
@@ -28,8 +28,8 @@ namespace sar.S7Siemens
 		public ushort dataBlock;
 		public uint startAddress;
 		public ushort length;
-		public ushort byteLength;
-
+		//public ushort byteLength;
+        public uint byteLength;
 		public ushort byteAdddress;
 		public ushort bitAddress;
 		public TransportType transportType;
